collector/requirements.txt
paho-mqtt>=2.0
supabase>=2.0
python-dotenv>=1.0
pytest>=8.0
